force int type on do_cloudfare option to correctly show the purge cloudflare cache link in admin menu

+++ inc/common/admin-bar.php
@@ -146,7 +146,7 @@ function rocket_admin_bar( $wp_admin_bar ) {
 		}
 
 		// Purge CloudFlare cache if CloudFlare is active.
-		if ( 1 === get_rocket_option( 'do_cloudflare', 0 ) ) {
+		if ( 1 === (int) get_rocket_option( 'do_cloudflare', 0 ) ) {
 			$action = 'rocket_purge_cloudflare';
 
 			$wp_admin_bar->add_menu( array(
